Move files to final architecture
