+++ inc/classes/class-plugin.php
@@ -119,7 +119,7 @@ class Plugin {
 			}
 		}
 
-		$this->container->addServiceProvider( 'WP_Rocket\Engine\Optimization\ServiceProvider' );
+		$this->container->addServiceProvider( 'WP_Rocket\Engine\CriticalPath\ServiceProvider' );
 
 		$this->container->addServiceProvider( 'WP_Rocket\Addon\ServiceProvider' );
 		$this->container->addServiceProvider( 'WP_Rocket\Engine\Preload\ServiceProvider' );
