+++ tests/Unit/inc/Engine/CriticalPath/CriticalCSS/cleanCriticalCss.php
@@ -1,19 +1,19 @@
 <?php
 
-namespace WP_Rocket\Tests\Unit\inc\Engine\Optimization\CSS\CriticalPath\CriticalCSS;
+namespace WP_Rocket\Tests\Unit\inc\Engine\CriticalPath\CriticalCSS;
 
 use Mockery;
 use Brain\Monkey\Functions;
 use WP_Rocket\Tests\Unit\FilesystemTestCase;
-use WP_Rocket\Engine\Optimization\CSS\CriticalPath\CriticalCSS;
-use WP_Rocket\Engine\Optimization\CSS\CriticalPath\CriticalCSSGeneration;
+use WP_Rocket\Engine\CriticalPath\CriticalCSS;
+use WP_Rocket\Engine\CriticalPath\CriticalCSSGeneration;
 
 /**
- * @covers \WP_Rocket\Engine\Optimization\CSS\CriticalPath\CriticalCSS::clean_critical_css
+ * @covers \WP_Rocket\Engine\CriticalPath\CriticalCSS::clean_critical_css
  * @group  CriticalCss
  */
 class Test_CleanCriticalCSS extends FilesystemTestCase {
-	protected $path_to_test_data = '/inc/Engine/Optimization/CSS/CriticalPath/CriticalCSS/cleanCriticalCss.php';
+	protected $path_to_test_data = '/inc/Engine/CriticalPath/CriticalCSS/cleanCriticalCss.php';
 
 	private $critical_css;
 	private $critical_css_path;
